+++ Terrain_Edtior/Terrain_Edtior.vcxproj
@@ -151,27 +151,27 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClCompile Include="glad.c" />
-    <ClCompile Include="Joystick.cpp" />
-    <ClCompile Include="Keyboard.cpp" />
+    <ClCompile Include="IO\Joystick.cpp" />
+    <ClCompile Include="IO\Keyboard.cpp" />
+    <ClCompile Include="IO\Mouse.cpp" />
+    <ClCompile Include="Libs\glad.c" />
+    <ClCompile Include="Libs\stb.cpp" />
     <ClCompile Include="main.cpp" />
-    <ClCompile Include="Mouse.cpp" />
-    <ClCompile Include="Shader.cpp" />
-    <ClCompile Include="stb.cpp" />
+    <ClCompile Include="Renderer\Shaders\Shader.cpp" />
   </ItemGroup>
   <ItemGroup>
-    <None Include="fragment_shader.glsl" />
-    <None Include="vertex_shader.glsl" />
+    <None Include="Assets\Shaders\fragment_shader.glsl" />
+    <None Include="Assets\Shaders\vertex_shader.glsl" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="Joystick.h" />
-    <ClInclude Include="Keyboard.h" />
-    <ClInclude Include="Mouse.h" />
-    <ClInclude Include="Shader.h" />
+    <ClInclude Include="IO\Joystick.h" />
+    <ClInclude Include="IO\Keyboard.h" />
+    <ClInclude Include="IO\Mouse.h" />
+    <ClInclude Include="Renderer\Shaders\Shader.h" />
   </ItemGroup>
   <ItemGroup>
-    <Image Include="smile_face.png" />
-    <Image Include="wall.jpg" />
+    <Image Include="Assets\Textures\smile_face.png" />
+    <Image Include="Assets\Textures\wall.jpg" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
